requirements.txt
Django==1.11.5
PyJWT==1.6.4
Pygments==2.2.0
argparse==1.2.1
certifi==2018.4.16
chardet==3.0.4
django-cors-headers==2.3.0
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
freesms==0.1.2
httplib2==0.11.3
idna==2.7
psycopg2==2.7.5
pytz==2018.4
redis==2.10.6
requests==2.19.1
sinchsms==1.0.4
six==1.11.0
twilio==5.7.0
urllib3==1.21.1
whitenoise==3.3.1
wsgiref==0.1.2
gunicorn==19.0.0
dj-database-url>=0.5.0
python-dateutil==2.5.1